Initial floating point mixer efforts

+++ libraries/libRAD/radmidi.cpp
@@ -34,7 +34,7 @@ bool OPLPlayer::loadPatches()
 }
 
 // ----------------------------------------------------------------------------
-void OPLPlayer::generate(int16_t *data, unsigned numSamples)
+void OPLPlayer::generate(float *data, unsigned numSamples)
 {
     unsigned int samp = 0;
 
@@ -127,7 +127,7 @@ void OPLPlayer::runSamples(unsigned count)
     // (i.e. when forcing a voice off, changing 4op flags, etc.)
     while (count--)
     {
-        std::pair<int16_t, int16_t> output;
+        std::pair<float, float> output;
         m_opl3->Sample(&output.first, &output.second);
         m_sampleFIFO.push(output);
     }
